authPrompt: use correct service name when answering preemptively
If the user hits enter before a question is asked, we try to answer
that question as soon as it comes in.

This doesn't work right now because we only answer the question if
there's a query service, but there won't ever be one until the
question is asked.

This commit corrects the code to it's intention, which is to answer
the first service that asks a question, as as soon as possible.