iconGrid: Implement hack for multiple line labels
Use the spacing and bottom padding of the iconGrid to allow
multiple line item labels. This way labels don't get cut off
after the first line and more information can be displayed.

It does work quite well but there's one visual inconsistency:
If the height of the label exceeds the available space, we
just assign the available height to the container. This leads to
inconsistent bottom paddings on some items because the text
lines of the label (obviously) have fixed heights.
This could only be fixed if there was a way to get the childs
natural height within a specified height limit. This way
the container would have the same height as the child inside
it and paddings would be consistent across all items.

This is certainly not the best way to implement multiple line
labels. Doing it correctly would probably require rethinking
the whole IconGrid class since it heavily depends on items
having static widths and heights right now.